+++ app/layouts/default.vue
@@ -17,5 +17,6 @@
     </main>
 
     <AppFooter />
+    <AppStatusBar />
   </div>
 </template>